+++ Sources/PDFGenerator+Image.swift
@@ -8,16 +8,16 @@
 
 extension PDFGenerator {
     
-    func drawImage(_ container: Container, image: UIImage, size: CGSize, caption: NSAttributedString, sizeFit: ImageSizeFit, calculatingMetrics: Bool) {
+    func drawImage(_ container: Container, image: UIImage, size: CGSize, caption: NSAttributedString, sizeFit: ImageSizeFit, calculatingMetrics: Bool) throws {
         var (imageSize, captionSize) = calculateImageCaptionSize(container, image: image, size: size, caption: caption, sizeFit: sizeFit)
         
-        let y: CGFloat = {
+        let y: CGFloat = try {
             switch container.normalize {
             case .headerLeft:
                 return headerHeight[container]!
             case .contentLeft:
                 if (contentHeight + imageSize.height + captionSize.height > contentSize.height || (sizeFit == .height && imageSize.height < size.height)) {
-                    generateNewPage(calculatingMetrics: calculatingMetrics)
+                    try generateNewPage(calculatingMetrics: calculatingMetrics)
                     
                     (imageSize, captionSize) = calculateImageCaptionSize(container, image: image, size: size, caption: caption, sizeFit: sizeFit)
                     
@@ -45,10 +45,10 @@ extension PDFGenerator {
         }()
         
         let frame = CGRect(x: x, y: y, width: imageSize.width, height: imageSize.height)
-        drawImage(container, image: image, frame: frame, caption: caption, calculatingMetrics: calculatingMetrics)
+        try drawImage(container, image: image, frame: frame, caption: caption, calculatingMetrics: calculatingMetrics)
     }
     
-    func drawImagesInRow(_ container: Container, images: [UIImage], captions: [NSAttributedString], spacing: CGFloat, calculatingMetrics: Bool) {
+    func drawImagesInRow(_ container: Container, images: [UIImage], captions: [NSAttributedString], spacing: CGFloat, calculatingMetrics: Bool) throws {
         assert(images.count > 0, "You need to provide at least one image!")
         
         let totalImagesWidth = contentSize.width - indentation[container.normalize]! - (CGFloat(images.count) - 1) * spacing
@@ -75,7 +75,7 @@ extension PDFGenerator {
         
         var y = contentHeight + maxHeaderHeight() + headerSpace
         if (contentHeight + maxHeight > contentSize.height) {
-            generateNewPage(calculatingMetrics: calculatingMetrics)
+            try generateNewPage(calculatingMetrics: calculatingMetrics)
             
             y = contentHeight + maxHeaderHeight() + headerSpace
             (imageSizes, maxHeight) = calculateImageCaptionSizes(images, captions)
@@ -87,7 +87,7 @@ extension PDFGenerator {
         for (index, image) in images.enumerated() {
             let imageSize = imageSizes[index]
             let caption = (captions.count > index) ? captions[index] : NSAttributedString()
-            drawImage(container, image: image, frame: CGRect(x: x, y: y, width: imageSize.width, height: imageSize.height), caption: caption, calculatingMetrics: calculatingMetrics)
+            try drawImage(container, image: image, frame: CGRect(x: x, y: y, width: imageSize.width, height: imageSize.height), caption: caption, calculatingMetrics: calculatingMetrics)
             
             x += imageSize.width + spacing
             indentation[container.normalize] = indentation[container.normalize]! + imageSize.width + spacing
@@ -98,28 +98,16 @@ extension PDFGenerator {
         contentHeight += maxHeight
     }
     
-    func drawImage(_ container: Container, image: UIImage, frame: CGRect, caption: NSAttributedString, calculatingMetrics: Bool) {
-        // resize
-        let resizeFactor = (3 * imageQuality > 1) ? 3 * imageQuality : 1
-        let resizeImageSize = CGSize(width: frame.size.width * resizeFactor, height: frame.size.height * resizeFactor)
+    func drawImage(_ container: Container, image: UIImage, frame: CGRect, caption: NSAttributedString, calculatingMetrics: Bool) throws {
+        let compressedImage = resizeAndCompressImage(image: image, frame: frame)
         
-        UIGraphicsBeginImageContext(resizeImageSize)
-        image.draw(in: CGRect(x:0, y:0, width: resizeImageSize.width, height: resizeImageSize.height))
-        var compressedImage = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        
-        // compression
-        if let image = compressedImage, let jpegData = UIImageJPEGRepresentation(image, imageQuality) {
-            compressedImage = UIImage(data: jpegData)
-        }
-        
-        if let resultImage = compressedImage {
-            // Don't render when calculating metrics
-            if !calculatingMetrics {
+        // Don't render when calculating metrics
+        if !calculatingMetrics {
+            if let resultImage = compressedImage {
                 resultImage.draw(in: frame)
+            } else {
+                image.draw(in: frame)
             }
-        } else {
-            image.draw(in: frame)
         }
         
         if container.isHeader {
@@ -131,7 +119,24 @@ extension PDFGenerator {
         }
         
         if caption.length > 0 {
-            drawAttributedText(container, text: caption, textMaxWidth: frame.size.width, calculatingMetrics: calculatingMetrics)
+            try drawAttributedText(container, text: caption, textMaxWidth: frame.size.width, calculatingMetrics: calculatingMetrics)
         }
+    }
+    
+    func resizeAndCompressImage(image: UIImage, frame: CGRect) -> UIImage? {
+        // resize
+        let resizeFactor = (3 * imageQuality > 1) ? 3 * imageQuality : 1
+        let resizeImageSize = CGSize(width: frame.size.width * resizeFactor, height: frame.size.height * resizeFactor)
+        
+        UIGraphicsBeginImageContext(resizeImageSize)
+        image.draw(in: CGRect(x: 0, y: 0, width: resizeImageSize.width, height: resizeImageSize.height))
+        var compressedImage = UIGraphicsGetImageFromCurrentImageContext()
+        UIGraphicsEndImageContext()
+        
+        // compression
+        if let image = compressedImage, let jpegData = UIImageJPEGRepresentation(image, imageQuality) {
+            compressedImage = UIImage(data: jpegData)
+        }
+        return compressedImage
     }
 }